Change escape mapping to C-space
It's slightly less ergonomic to type, but it avoids the weird thing
where trying to enter a literal j and the j doesn't show up until you
type another character that is not k. Unfortunately, helix does not seem
to have a way to require you type the subsequent characters within a
certain timeout.