Switch back to Hyprland
I *need* a good tiling window manager. While the KDE desktop is pretty
and complete, the kwinscript implementation of tiling leaves a lot to be
desired. I tried Cosmic, but it is still too buggy (alpha software, so
not surprising.) Hyprland may take more effort to set up/maintain, but
it is beautiful, configurable in all the right ways, and pretty stable.